[Update] Updated Main class to use ExchangeRateFileReaderHandler, moved some utils methods which were responsibility for data validations from Main class to InputValidator class
